fix: close delivery verification gaps
