+++ src/clients/doctor.rs
@@ -111,7 +111,7 @@ impl ClientManager {
 const MINIMUM_CLAUDE_GATEWAY_VERSION: (u64, u64, u64) = (2, 1, 129);
 
 /// Claude Code began accepting gateway-discovered model ids in 2.1.129.
-pub(crate) fn claude_gateway_version_supported(output: &str) -> bool {
+fn claude_gateway_version_supported(output: &str) -> bool {
     output
         .split(|character: char| !(character.is_ascii_digit() || character == '.'))
         .find_map(|part| {
@@ -126,6 +126,7 @@ pub(crate) fn claude_gateway_version_supported(output: &str) -> bool {
 }
 
 /// Fail with an actionable diagnostic when an installed Claude is too old.
+#[allow(clippy::redundant_pub_crate)]
 pub(crate) fn require_claude_gateway_version() -> Result<(), ClientError> {
     let Ok(output) = std::process::Command::new("claude")
         .arg("--version")
