+++ src/model_routing_tests.rs
@@ -11,6 +11,7 @@ use std::sync::Arc;
 use tempfile::tempdir;
 use tower::ServiceExt;
 
+#[allow(clippy::redundant_pub_crate)]
 pub(crate) fn auto_state(readers: Vec<SubscriptionReader>, data_dir: &std::path::Path) -> AppState {
     AppState {
         client: reqwest::Client::new(),
@@ -53,6 +54,7 @@ pub(crate) fn auto_state(readers: Vec<SubscriptionReader>, data_dir: &std::path:
     }
 }
 
+#[allow(clippy::redundant_pub_crate)]
 pub(crate) fn bound_client_token(
     state: &AppState,
     client: crate::clients::ClientKind,
@@ -799,243 +801,4 @@ fn an_unavailable_credential_is_named_in_the_routing_error() {
         message.contains("missing or rejected upstream"),
         "{message}"
     );
-}
-
-/// Add a stored OpenAI-compatible provider declaring `models`.
-fn store_provider(state: &AppState, name: &str, models: &[&str]) {
-    state
-        .provider_store
-        .upsert(crate::providers::ProviderUpsert {
-            name: name.to_string(),
-            kind: None,
-            base_url: "https://provider.example/v1".to_string(),
-            default_model: models.first().map(|model| (*model).to_string()),
-            models: Some(models.iter().map(|model| (*model).to_string()).collect()),
-            api_key: Some("provider-key".to_string()),
-            api_key_env: None,
-            encrypted_api_key: None,
-            enabled: Some(true),
-            subscriber_id: None,
-            acknowledge_intermediary_risk: None,
-            acknowledge_unsupported_clients: None,
-        })
-        .expect("store the provider");
-}
-
-/// The bug in issue #260: a stored provider was reachable only by pinning
-/// `UPSTREAM_PROVIDER`, which pins the whole deployment — so one router could
-/// serve vendor subscriptions or a local endpoint, never both.
-#[tokio::test]
-async fn a_stored_providers_declared_model_routes_in_automatic_mode() {
-    let data_dir = tempfile::tempdir().expect("data dir");
-    let state = auto_state(Vec::new(), data_dir.path());
-    store_provider(&state, "formal-ai", &["formal-ai-mini"]);
-
-    let routed =
-        crate::model_routing::route_state(&state, &serde_json::json!({"model": "formal-ai-mini"}))
-            .await
-            .expect("a declared model must route");
-
-    assert_eq!(routed.upstream_provider, UpstreamProvider::OpenAICompatible);
-    assert_eq!(routed.openai_compatible.provider_name, "formal-ai");
-    assert_eq!(routed.bridge_model.as_deref(), Some("formal-ai-mini"));
-    // The deployment itself is untouched: this routed one request.
-    assert_eq!(state.upstream_provider, UpstreamProvider::Auto);
-}
-
-/// A declared model appears in `/v1/models`, so one token reaches every model
-/// the router can serve rather than only the discovered subscriptions.
-#[tokio::test]
-async fn declared_models_are_listed_alongside_subscription_catalogs() {
-    let data_dir = tempfile::tempdir().expect("data dir");
-    let state = auto_state(Vec::new(), data_dir.path());
-    store_provider(&state, "formal-ai", &["formal-ai-mini", "formal-ai-large"]);
-
-    let mut catalog = crate::model_routing::model_catalog(&[], &state.model_catalogs);
-    crate::model_routing::append_stored_provider_models(&state, &mut catalog);
-
-    let ids: Vec<&str> = catalog["data"]
-        .as_array()
-        .expect("a data array")
-        .iter()
-        .filter_map(|entry| entry["id"].as_str())
-        .collect();
-    assert!(ids.contains(&"formal-ai-mini"), "{ids:?}");
-    assert!(ids.contains(&"formal-ai-large"), "{ids:?}");
-}
-
-/// A model declared by two stored providers is refused rather than resolved by
-/// declaration order — the rule subscriptions already follow.
-#[tokio::test]
-async fn a_model_declared_twice_is_ambiguous_until_qualified() {
-    let data_dir = tempfile::tempdir().expect("data dir");
-    let state = auto_state(Vec::new(), data_dir.path());
-    store_provider(&state, "alpha", &["shared-model"]);
-    store_provider(&state, "beta", &["shared-model"]);
-
-    // Matched rather than `expect_err`: `AppState` holds credentials and so
-    // deliberately does not implement `Debug`.
-    let Err(error) =
-        crate::model_routing::route_state(&state, &serde_json::json!({"model": "shared-model"}))
-            .await
-    else {
-        panic!("an ambiguous name must be refused");
-    };
-    assert!(
-        matches!(error, crate::model_routing::ModelRouteError::Ambiguous(_)),
-        "{error:?}"
-    );
-
-    // Naming the provider resolves it.
-    let routed = crate::model_routing::route_state(
-        &state,
-        &serde_json::json!({"model": "beta/shared-model"}),
-    )
-    .await
-    .expect("a qualified name is unambiguous");
-    assert_eq!(routed.openai_compatible.provider_name, "beta");
-    assert_eq!(routed.bridge_model.as_deref(), Some("shared-model"));
-}
-
-/// A disabled provider advertises nothing, so disabling one takes its models
-/// out of both the catalog and the routing table.
-#[tokio::test]
-async fn a_disabled_provider_advertises_nothing() {
-    let data_dir = tempfile::tempdir().expect("data dir");
-    let state = auto_state(Vec::new(), data_dir.path());
-    store_provider(&state, "formal-ai", &["formal-ai-mini"]);
-    state
-        .provider_store
-        .upsert(crate::providers::ProviderUpsert {
-            name: "formal-ai".to_string(),
-            kind: None,
-            base_url: "https://provider.example/v1".to_string(),
-            default_model: None,
-            models: Some(vec!["formal-ai-mini".to_string()]),
-            api_key: None,
-            api_key_env: None,
-            encrypted_api_key: None,
-            enabled: Some(false),
-            subscriber_id: None,
-            acknowledge_intermediary_risk: None,
-            acknowledge_unsupported_clients: None,
-        })
-        .expect("disable the provider");
-
-    let Err(error) =
-        crate::model_routing::route_state(&state, &serde_json::json!({"model": "formal-ai-mini"}))
-            .await
-    else {
-        panic!("a disabled provider must not route");
-    };
-    assert!(
-        matches!(error, crate::model_routing::ModelRouteError::NotFound(_)),
-        "{error:?}"
-    );
-}
-
-/// A qualified name that the provider does not advertise is an error naming
-/// the provider, rather than a silent fall through to a subscription.
-#[tokio::test]
-async fn a_qualified_name_the_provider_lacks_is_reported() {
-    let data_dir = tempfile::tempdir().expect("data dir");
-    let state = auto_state(Vec::new(), data_dir.path());
-    store_provider(&state, "formal-ai", &["formal-ai-mini"]);
-
-    let Err(error) = crate::model_routing::route_state(
-        &state,
-        &serde_json::json!({"model": "formal-ai/not-declared"}),
-    )
-    .await
-    else {
-        panic!("an undeclared qualified model must be refused");
-    };
-
-    assert!(
-        matches!(error, crate::model_routing::ModelRouteError::NotFound(_)),
-        "{error:?}"
-    );
-    assert!(format!("{error:?}").contains("formal-ai"), "{error:?}");
-}
-
-/// A qualified name for a provider that does not exist falls through to
-/// ordinary routing rather than being treated as a provider reference.
-#[tokio::test]
-async fn an_unknown_provider_prefix_is_not_a_provider_reference() {
-    let data_dir = tempfile::tempdir().expect("data dir");
-    let state = auto_state(Vec::new(), data_dir.path());
-
-    let Err(error) =
-        crate::model_routing::route_state(&state, &serde_json::json!({"model": "nobody/model"}))
-            .await
-    else {
-        panic!("nothing advertises this model");
-    };
-
-    assert!(
-        matches!(error, crate::model_routing::ModelRouteError::NotFound(_)),
-        "{error:?}"
-    );
-}
-
-/// A stored model whose id collides with a subscription's is listed in its
-/// qualified form, so both stay reachable and the bare id stays ambiguous.
-#[tokio::test]
-async fn a_colliding_declared_model_is_listed_qualified() {
-    let data_dir = tempfile::tempdir().expect("data dir");
-    let state = auto_state(Vec::new(), data_dir.path());
-    store_provider(&state, "formal-ai", &["shared-id"]);
-
-    let mut catalog = serde_json::json!({
-        "object": "list",
-        "data": [{"id": "shared-id", "object": "model", "owned_by": "anthropic"}]
-    });
-    crate::model_routing::append_stored_provider_models(&state, &mut catalog);
-
-    let ids: Vec<&str> = catalog["data"]
-        .as_array()
-        .expect("data")
-        .iter()
-        .filter_map(|entry| entry["id"].as_str())
-        .collect();
-    assert!(
-        ids.contains(&"shared-id"),
-        "the subscription keeps its id: {ids:?}"
-    );
-    assert!(
-        ids.contains(&"formal-ai/shared-id"),
-        "the stored provider is reachable by its qualified name: {ids:?}"
-    );
-}
-
-/// A request with no model is refused before any provider is consulted.
-#[tokio::test]
-async fn a_request_without_a_model_is_refused() {
-    let data_dir = tempfile::tempdir().expect("data dir");
-    let state = auto_state(Vec::new(), data_dir.path());
-
-    let Err(error) = crate::model_routing::route_state(&state, &serde_json::json!({})).await else {
-        panic!("a model is required in automatic mode");
-    };
-
-    assert!(
-        matches!(error, crate::model_routing::ModelRouteError::ModelRequired),
-        "{error:?}"
-    );
-}
-
-#[test]
-fn automatic_routing_errors_never_expose_catalog_bodies_accounts_or_paths() {
-    let catalogs = ModelCatalogCache::new();
-    let sentinel = "vendor-body account-secret /private/credentials/codex.json";
-    catalogs.record_failure(SubscriptionProvider::Codex, sentinel, true);
-
-    let error = available_provider_for_model("gpt-secret", &[], &catalogs)
-        .expect_err("a failed catalog is not routable")
-        .to_string();
-
-    assert!(error.contains("codex"));
-    assert!(!error.contains("vendor-body"), "{error}");
-    assert!(!error.contains("account-secret"), "{error}");
-    assert!(!error.contains("/private/credentials"), "{error}");
 }
